on ia64 native environment, we directly call gcc/g++ for assembling and linking for C and C++ source. we still directly use ld to link fortran objects
git-svn-id: http://svn.open64.net/svnroot/open64@316 ead328b1-e903-0410-a19d-eacee4b031f6

+++ projects/pathscale-2.2.1/kpro64/driver/main.c
@@ -353,7 +353,7 @@ main (int argc, char *argv[])
 		    "in the\n");
 	    fprintf(stderr, "%.*s/share/doc/pathscale-compilers-" 
 	    	    PSC_FULL_VERSION "/LEGAL.pdf file.\n",
-		    strlen(exe_dir) - 4, exe_dir);
+		    (int)(strlen(exe_dir) - 4), exe_dir);
 	}
 	if (show_search_path) {
 		print_search_path();
@@ -1089,7 +1089,7 @@ print_search_path ()
 	char *gcc_lib_ptr;
 	int buflen;
 	
-	printf ("install: %.*s\n", strlen(exe_dir) - 4, exe_dir);
+	printf ("install: %.*s\n", (int)(strlen(exe_dir) - 4), exe_dir);
 	printf ("programs: %s:%s\n", exe_dir, get_phase_dir (P_be));
 	
 	if (abi == ABI_N32) {
